+++ BrillouinAcquisition/src/phase.h
@@ -77,11 +77,11 @@ private:
 	}
 
 	std::vector<int> createMask(int dim_x, int dim_y, double maskRadius) {
-		std::vector<int> mask(dim_x * dim_y, 0);
+		std::vector<int> mask((size_t)dim_x * dim_y, 0);
 		for (int x{ (int)round(dim_x / 2.0 - m_maskRadius) }; x < round(dim_x / 2.0 + m_maskRadius); x++) {
 			for (int y{ (int)round(dim_y / 2.0 - m_maskRadius) }; y < round(dim_y / 2.0 + m_maskRadius); y++) {
 				if (sqrt(pow((x - dim_x/2.0), 2) + pow(y - dim_y/2.0, 2)) <= m_maskRadius) {
-					mask[x + dim_x * y] = 1;
+					mask[x + (size_t)dim_x * y] = 1;
 				}
 			}
 		}
@@ -173,7 +173,7 @@ public:
 
 		// Calculate magnitude of background image to find the indices of the largest element
 		std::vector<double> background;
-		background.resize(dim_x * dim_y);
+		background.resize((size_t)dim_x * dim_y);
 		for (int i{ 0 }; i < dim_x * dim_y; i++) {
 			background[i] = pow(m_out_FFT[i][0], 2) + pow(m_out_FFT[i][1], 2);
 		}
@@ -208,7 +208,7 @@ public:
 
 		// Calculate the absolute value
 		for (int i{ 0 }; i < dim_x * dim_y; i++) {
-			(*spectrum)[i] = log10(sqrt(pow(m_out_FFT[i][0], 2) + pow(m_out_FFT[i][1], 2)) / (dim_x * dim_y));
+			(*spectrum)[i] = log10(sqrt(pow(m_out_FFT[i][0], 2) + pow(m_out_FFT[i][1], 2)) / ((size_t)dim_x * dim_y));
 		}
 
 		fftshift(&(*spectrum)[0], dim_x, dim_y);
@@ -254,7 +254,7 @@ public:
 		int dim_x_new{ dim_x / 3 };
 		int dim_y_new{ dim_y / 3 };
 		std::vector<float> phase_lowRes;
-		phase_lowRes.resize(dim_x_new * dim_y_new);
+		phase_lowRes.resize((size_t)dim_x_new * dim_y_new);
 		m_xsample->resample(&(*phase)[0], &phase_lowRes[0], dim_x, dim_y, dim_x_new, dim_y_new, RESAMPLE_MODE::NEAREST);
 
 		std::vector<float> phaseUnwrapped = phase_lowRes;
